+++ package.json
@@ -26,11 +26,11 @@
     "@types/node": "^22.15.21",
     "esbuild": "0.23.1",
     "htmlparser2": "^10.0.0",
-    "lit": "^3.2.1",
     "prettier": "^3.5.3",
     "storybook": "^8.4.7",
     "typescript": "^5.8.3",
-    "vite": "^5.4.1"
+    "vite": "^5.4.1",
+    "vite-plugin-dts": "^4.5.4"
   },
   "dependencies": {
     "@glidejs/glide": "^3.7.1",
