+++ src/common/theme.ts
@@ -23,97 +23,6 @@ export function getSuccessAnimation(
   return checkmarkIcon(theme, width, height);
 }
 
-export function getFollowButtonStyles(theme: Theme, isLoading: boolean) {
-  let variables = ``;
-
-  if (theme === 'dark') {
-    variables = `
-          --nstrc-follow-btn-background: var(--nstrc-follow-btn-background-dark);
-          --nstrc-follow-btn-hover-background: var(--nstrc-follow-btn-hover-background-dark);
-    
-          --nstrc-follow-btn-text-color: var(--nstrc-follow-btn-text-color-dark);
-          --nstrc-follow-btn-border: var(--nstrc-follow-btn-border-dark);
-      `;
-  } else {
-    variables = `
-          --nstrc-follow-btn-background: var(--nstrc-follow-btn-background-light);
-          --nstrc-follow-btn-hover-background: var(--nstrc-follow-btn-hover-background-light);
-    
-          --nstrc-follow-btn-text-color: var(--nstrc-follow-btn-text-color-light);
-          --nstrc-follow-btn-border: var(--nstrc-follow-btn-border-light);
-      `;
-  }
-
-  return `
-        <style>
-          :host {
-            ${variables}
-
-            --nstrc-follow-btn-padding: 10px 16px;
-            --nstrc-follow-btn-font-size: 16px;
-            --nstrc-follow-btn-background-dark: #000000;
-            --nstrc-follow-btn-background-light: #FFFFFF;
-            --nstrc-follow-btn-hover-background-dark: #222222;
-            --nstrc-follow-btn-hover-background-light: #F9F9F9;
-            --nstrc-follow-btn-border-dark: none;
-            --nstrc-follow-btn-border-light: 1px solid #DDDDDD;
-            --nstrc-follow-btn-text-color-dark: #FFFFFF;
-            --nstrc-follow-btn-text-color-light: #000000;
-            --nstrc-follow-btn-border-radius: 8px;
-            --nstrc-follow-btn-error-font-size: 12px;
-            --nstrc-follow-btn-error-line-height: 1em;
-            --nstrc-follow-btn-error-max-width: 250px;
-            --nstrc-follow-btn-horizontal-alignment: start;
-            --nstrc-follow-btn-min-height: 47px;
-          }
-
-          .nostr-follow-button-container {
-            display: flex;
-            flex-direction: column;
-            font-family: Inter,sans-serif;
-            flex-direction: column;
-            gap: 8px;
-            width: fit-content;
-          }
-
-          .nostr-follow-button-wrapper {
-            display: flex;
-            justify-content: var(--nstrc-follow-btn-horizontal-alignment);
-          }
-    
-          .nostr-follow-button {
-            display: flex;
-            align-items: center;
-            gap: 12px;
-            border-radius: var(--nstrc-follow-btn-border-radius);
-            background-color: var(--nstrc-follow-btn-background);
-            cursor: pointer;
-
-            min-height: var(--nstrc-follow-btn-min-height);
-
-            border: var(--nstrc-follow-btn-border);
-
-            padding: var(--nstrc-follow-btn-padding);
-            font-size: var(--nstrc-follow-btn-font-size);
-            color: var(--nstrc-follow-btn-text-color);
-
-            ${isLoading ? 'pointer-events: none; user-select: none; background-color: var(--nstrc-follow-btn-hover-background);' : ''}
-          }
-
-          .nostr-follow-button:hover {
-            background-color: var(--nstrc-follow-btn-hover-background);
-          }
-
-          .nostr-follow-button-error small {
-            justify-content: flex-end;
-            color: red;
-            font-size: var(--nstrc-follow-btn-error-font-size);
-            line-height: var(--nstrc-follow-btn-error-line-height);
-            max-width: var(--nstrc-follow-btn-error-max-width);
-          }
-        </style>
-    `;
-}
 
 // Function for NostrPost styles (legacy/innerHTML)
 export function getPostStylesLegacy(theme: Theme): string {
@@ -383,7 +292,7 @@ export const getProfileBadgeStylesLegacy = (theme: Theme) => {
       background-color: var(--nstrc-profile-badge-background-light);
       color: var(--nstrc-profile-badge-text-color-light);
       cursor: pointer;
-      font-family: sans-serif; /* Added default font */
+      font-family: Inter, sans-serif; /* Added default font */
       max-width: 300px; /* Added max-width */
       overflow: hidden; /* Prevent content overflow */
       text-overflow: ellipsis; /* Add ellipsis for overflow */
