+++ stories/NostrFollowButton.stories.tsx
@@ -165,7 +165,7 @@ const generateCode = args => {
   const bundleScript = `<script type="module" src="/nostr-components.es.js"></script>`;
 
   const cssVariables = CSS_VARIABLES.map(cssVariable => {
-    console.log(args[cssVariable.variable]);
+    // console.log(args[cssVariable.variable]);
     return args[cssVariable.variable]
       ? `${cssVariable.variable}: ${args[cssVariable.variable]} !important;`
       : '';
